+++ docker/docker.go
@@ -68,7 +68,7 @@ func mhzToVCPU(cpuInMhz int) float64 {
 // RunContainer goes through the process of setting constraints,
 // specifying image name and cmd. It starts a container and posts
 // log update every gistUpdateDuration.
-func RunContainer(depReq models.DeploymentRequest, createdGist *github.Gist, resCh chan<- models.DeploymentResponse) {
+func RunContainer(depReq models.DeploymentRequest, createdGist *github.Gist, resCh chan<- models.DeploymentResponse, servicePK uint) {
 	log.Println("Entering RunContainer")
 	machine_type := depReq.Params.MachineType
 	gpuOpts := opts.GpuOpts{}
@@ -138,11 +138,12 @@ func RunContainer(depReq models.DeploymentRequest, createdGist *github.Gist, res
 		panic(res.Error)
 	}
 
-	// Update db
+	// Update db - find the service based on primary key and update container id
 	var service models.Services
-	service.ContainerID = resp.ID
-	service.ImageID = depReq.Params.ImageID
-	service.ServiceName = depReq.Params.ImageID
+	res = db.DB.Model(&service).Updates(models.Services{ContainerID: resp.ID})
+	if res.Error != nil {
+		panic(res.Error)
+	}
 
 	var resourceRequirements models.ServiceResourceRequirements
 	resourceRequirements.CPU = depReq.Constraints.CPU
@@ -154,10 +155,7 @@ func RunContainer(depReq models.DeploymentRequest, createdGist *github.Gist, res
 	}
 
 	service.ResourceRequirements = int(resourceRequirements.ID)
-	result = db.DB.Create(&service)
-	if result.Error != nil {
-		panic(result.Error)
-	}
+	// TODO: Update service based on passed pk
 
 	telemetry.CalcFreeResources()
 	libp2p.UpdateDHT()
@@ -311,25 +309,39 @@ func cpuUsage(cpu float64, maxCPU float64) float64 {
 
 // HandleDeployment does following docker based actions in the sequence:
 // Pull image, run container, get logs, delete container, send log to the requester
-func HandleDeployment(depReq models.DeploymentRequest, depRes models.DeploymentResponse) models.DeploymentResponse {
+func HandleDeployment(depReq models.DeploymentRequest) models.DeploymentResponse {
 	// Pull the image
 	imageName := depReq.Params.ImageID
 	PullImage(imageName)
 
+	// create a service and pass the primary key to the RunContainer to update ContainerID
+	var service models.Services
+	service.ImageID = depReq.Params.ImageID
+	service.ServiceName = depReq.Params.ImageID
+	service.JobStatus = "running"
+	service.JobDuration = 5           // these are dummy data, implementation pending
+	service.EstimatedJobDuration = 10 // these are dummy data, implementation pending
+
+	// create gist here and pass it to RunContainer to update logs
 	createdGist, _, err := createGist()
 	if err != nil {
 		log.Panicln(err)
 	}
 
+	service.LogURL = *createdGist.HTMLURL
+	// Save the service with logs
+	result := db.DB.Create(&service)
+	if result.Error != nil {
+		panic(result.Error)
+	}
 	resCh := make(chan models.DeploymentResponse)
 
 	// Run the container.
-	go RunContainer(depReq, createdGist, resCh)
+	go RunContainer(depReq, createdGist, resCh, service.ID)
 
 	res := <-resCh
 
 	// Send back *createdGist.HTMLURL
-	depRes.Content = *createdGist.HTMLURL
-	depRes.Success = res.Success
-	return depRes
+	res.Content = *createdGist.HTMLURL
+	return res
 }